Remove some more unnecessary allocations

+++ community/kernel/src/main/java/org/neo4j/kernel/impl/api/security/OverriddenAccessMode.java
@@ -72,7 +72,7 @@ public class OverriddenAccessMode extends WrappedAccessMode
     }
 
     @Override
-    public boolean allowsTraverseLabels( int... labels )
+    public boolean allowsTraverseLabels( long... labels )
     {
         return wrapping.allowsTraverseLabels( labels );
     }
