+++ community/kernel/src/main/java/org/neo4j/kernel/impl/api/security/RestrictedAccessMode.java
@@ -72,7 +72,7 @@ public class RestrictedAccessMode extends WrappedAccessMode
     }
 
     @Override
-    public boolean allowsTraverseLabels( int... labels )
+    public boolean allowsTraverseLabels( long... labels )
     {
         return original.allowsTraverseLabels( labels ) && wrapping.allowsTraverseLabels( labels );
     }
